Update functions to replace parseInt w/ parseFloat
Add support for basic operations that include
decimal points

+++ javascript.js
@@ -114,11 +114,11 @@ function handleInput(value) {
 }
 
 function add(a, b) {
-	return parseInt(a) + parseInt(b);
+	return parseFloat(a) + parseFloat(b);
 };
 
 function subtract(a, b) {
-	return parseInt(a) - parseInt(b);
+	return parseFloat(a) - parseFloat(b);
 };
 
 function sum(arr) {
@@ -132,11 +132,11 @@ function multiplyArr(arr) {
 };
 
 function multiply(a, b) {
-    return parseInt(a) * parseInt(b);
+    return parseFloat(a) * parseFloat(b);
 }
 
 function divide(a, b) {
-    return parseInt(a) / parseInt(b);
+    return parseFloat(a) / parseFloat(b);
 }
 
 function displayMem() {
